+++ logprep/framework/pipeline_manager.py
@@ -17,16 +17,6 @@ from logprep.metrics.metrics import CounterMetric
 from logprep.util.configuration import Configuration
 
 
-def logger_process(queue: multiprocessing.queues.Queue, logger: logging.Logger):
-    """Process log messages from a queue."""
-
-    while True:
-        message = queue.get()
-        if message is None:
-            break
-        logger.handle(message)
-
-
 class PipelineManager:
     """Manage pipelines via multi-processing."""
 
@@ -60,9 +50,8 @@ class PipelineManager:
 
     def __init__(self, configuration: Configuration):
         self.metrics = self.Metrics(labels={"component": "manager"})
-        self._logger = logging.getLogger("Logprep PipelineManager")
+        self._logger = logging.getLogger("Manager")
         if multiprocessing.current_process().name == "MainProcess":
-            self._start_multiprocess_logger()
             self._set_http_input_queue(configuration)
         self._pipelines: list[multiprocessing.Process] = []
         self._configuration = configuration
@@ -85,25 +74,6 @@ class PipelineManager:
             return
         message_backlog_size = input_config.get("message_backlog_size", 15000)
         HttpConnector.messages = multiprocessing.Queue(maxsize=message_backlog_size)
-
-    def _start_multiprocess_logger(self):
-        self.log_queue = multiprocessing.Queue(-1)
-        self._log_process = multiprocessing.Process(
-            target=logger_process, args=(self.log_queue, self._logger), daemon=True
-        )
-        self._log_process.start()
-
-    def get_count(self) -> int:
-        """Get the pipeline count.
-
-        Parameters
-        ----------
-        count : int
-           The pipeline count will be incrementally changed until it reaches this value.
-
-        """
-        self._logger.debug(f"Getting pipeline count: {len(self._pipelines)}")
-        return len(self._pipelines)
 
     def set_count(self, count: int):
         """Set the pipeline count.
@@ -161,9 +131,6 @@ class PipelineManager:
         self._decrease_to_count(0)
         if self.prometheus_exporter:
             self.prometheus_exporter.cleanup_prometheus_multiprocess_dir()
-        self.log_queue.put(None)  # signal the logger process to stop
-        self._log_process.join()
-        self.log_queue.close()
 
     def restart(self):
         """Restarts all pipelines"""
@@ -175,12 +142,7 @@ class PipelineManager:
             self.prometheus_exporter.run()
 
     def _create_pipeline(self, index) -> multiprocessing.Process:
-        pipeline = Pipeline(
-            pipeline_index=index,
-            config=self._configuration,
-            log_queue=self.log_queue,
-            lock=self._lock,
-        )
+        pipeline = Pipeline(pipeline_index=index, config=self._configuration, lock=self._lock)
         self._logger.info("Created new pipeline")
         process = multiprocessing.Process(target=pipeline.run, daemon=True)
         process.stop = pipeline.stop